Support CAP_DAC_READ_SEARCH capability
This allows you to read log files even if you would not ordinarily have
permissions to read them. This is useful if you do not control ownership
of the log files but still wish to ingest them. One use case is libvirt
qemu domain logs where there are no config options to set the ownership.

The container will detect if this capability is present on startup and
configure fluentd accordingly. It is possible to opt out of this
functionality by not giving the container the CAP_DAC_READ_SEARCH
capability, in which case, it should function as before. When using
CAP_DAC_READ_SEARCH care should be taken not to mount sensitive files
into the container (i.e only mount in files you don't mind it reading
e.g /var/log/kolla)

See: https://docs.fluentd.org/deployment/linux-capability
Change-Id: I7149f5ae3d137407eea9e1a81ca7ddfa8f439c96
Closes-Bug: #2049386