Support collecting libvirt qemu domain logs
This change causes the logs to be collected in /var/log/kolla.
Unfortuantly, it was not possible to change the log directory via a
config option, so we need to symlink the hardcoded path to a directory
under /var/log/kolla. The consequence of this is that the logs are
collected in the kolla logs volume where other programs, such as
fluentd, can read them.

Change-Id: If10006bf7f037a44b9b664b70ad423f7c6c08f45
Partial-Bug: #2049389