Backup from active mariadb server
Proxysql reports its version to MySQL clients,
and that's why the backup fails when ProxySQL used.

This patch modifies the script that handles backups, so that
instead of backing up through the load balancer, it just
connects through the load balancer, retrieves the address,
and then starts backing up directly from that server.

Closes-Bug: #2080818
Change-Id: Ibf7e40e8b059d733e114963022df06180249c650